Adding comments to app.tsx and refactoring file handling
